+++ src/components/organisims/Header/index.ts
@@ -1,0 +1,3 @@
+import { Header } from './header'
+
+export { Header }
